seperate card history component
